+++ src/services/HotelService/Hotel.Application/DTOs/RoomDTOs/UpdateRoomDto.cs
@@ -8,10 +8,10 @@ public record UpdateRoomDto
     public int TotalRoomCount { get; init; }
     public string RoomType { get; init; } 
     public int MaxOccupancy { get; init; }
-    public decimal Size { get; init; }
-    public decimal BasePrice { get; init; }
-    public decimal WeekendPriceMultiplier { get; init; }
-    public decimal SeasonPriceMultiplier { get; init; }
+    public int Size { get; init; }
+    public int BasePrice { get; init; }
+    public int WeekendPriceMultiplier { get; init; }
+    public int SeasonPriceMultiplier { get; init; }
 
 
     public bool HasWifi { get; init; }
